+++ archive/internal/gz_archiver.py
@@ -1,5 +1,6 @@
 import gzip
 import os
+from miscellaneous import humanSize
 from properties import *
 from logger import logger
 
@@ -20,7 +21,8 @@ def compress(targetPath:str, sch:dict) -> str:
 
     tfile.close()
     zfile.close()
-    logger.info('compress finished with success!')
+    zsize = os.path.getsize(zipPath)
+    logger.info(f'compress finished with success! compressed size: {humanSize(zsize)}')
     return os.path.basename(zipPath)